chore: move VecHeapAdapterInterface
