+++ vm/src/arch/integration_api.rs
@@ -390,6 +390,31 @@ impl<
     type ProcessedInstruction = PI;
 }
 
+#[derive(Clone)]
+pub struct VecHeapAdapterInterface<
+    T,
+    const R: usize,
+    const NUM_READS: usize,
+    const NUM_WRITES: usize,
+    const READ_SIZE: usize,
+    const WRITE_SIZE: usize,
+>(PhantomData<T>);
+
+impl<
+        T,
+        const R: usize,
+        const NUM_READS: usize,
+        const NUM_WRITES: usize,
+        const READ_SIZE: usize,
+        const WRITE_SIZE: usize,
+    > VmAdapterInterface<T>
+    for VecHeapAdapterInterface<T, R, NUM_READS, NUM_WRITES, READ_SIZE, WRITE_SIZE>
+{
+    type Reads = [[[T; READ_SIZE]; NUM_READS]; R];
+    type Writes = [[T; WRITE_SIZE]; NUM_WRITES];
+    type ProcessedInstruction = MinimalInstruction<T>;
+}
+
 /// Similar to `BasicAdapterInterface`, but it flattens the reads and writes into a single flat array for each
 pub struct FlatInterface<T, PI, const READ_CELLS: usize, const WRITE_CELLS: usize>(
     PhantomData<T>,
@@ -427,6 +452,112 @@ pub struct DynArray<T>(pub Vec<T>);
 
 mod conversions {
     use super::*;
+
+    // AdapterAirContext: VecHeapAdapterInterface -> DynInterface
+    impl<
+            T,
+            const R: usize,
+            const NUM_READS: usize,
+            const NUM_WRITES: usize,
+            const READ_SIZE: usize,
+            const WRITE_SIZE: usize,
+        >
+        From<
+            AdapterAirContext<
+                T,
+                VecHeapAdapterInterface<T, R, NUM_READS, NUM_WRITES, READ_SIZE, WRITE_SIZE>,
+            >,
+        > for AdapterAirContext<T, DynAdapterInterface<T>>
+    {
+        fn from(
+            ctx: AdapterAirContext<
+                T,
+                VecHeapAdapterInterface<T, R, NUM_READS, NUM_WRITES, READ_SIZE, WRITE_SIZE>,
+            >,
+        ) -> Self {
+            AdapterAirContext {
+                to_pc: ctx.to_pc,
+                reads: ctx.reads.into(),
+                writes: ctx.writes.into(),
+                instruction: ctx.instruction.into(),
+            }
+        }
+    }
+
+    // AdapterRuntimeContext: VecHeapAdapterInterface -> DynInterface
+    impl<
+            T,
+            const R: usize,
+            const NUM_READS: usize,
+            const NUM_WRITES: usize,
+            const READ_SIZE: usize,
+            const WRITE_SIZE: usize,
+        >
+        From<
+            AdapterRuntimeContext<
+                T,
+                VecHeapAdapterInterface<T, R, NUM_READS, NUM_WRITES, READ_SIZE, WRITE_SIZE>,
+            >,
+        > for AdapterRuntimeContext<T, DynAdapterInterface<T>>
+    {
+        fn from(
+            ctx: AdapterRuntimeContext<
+                T,
+                VecHeapAdapterInterface<T, R, NUM_READS, NUM_WRITES, READ_SIZE, WRITE_SIZE>,
+            >,
+        ) -> Self {
+            AdapterRuntimeContext {
+                to_pc: ctx.to_pc,
+                writes: ctx.writes.into(),
+            }
+        }
+    }
+
+    // AdapterAirContext: DynInterface -> VecHeapAdapterInterface
+    impl<
+            T,
+            const R: usize,
+            const NUM_READS: usize,
+            const NUM_WRITES: usize,
+            const READ_SIZE: usize,
+            const WRITE_SIZE: usize,
+        > From<AdapterAirContext<T, DynAdapterInterface<T>>>
+        for AdapterAirContext<
+            T,
+            VecHeapAdapterInterface<T, R, NUM_READS, NUM_WRITES, READ_SIZE, WRITE_SIZE>,
+        >
+    {
+        fn from(ctx: AdapterAirContext<T, DynAdapterInterface<T>>) -> Self {
+            AdapterAirContext {
+                to_pc: ctx.to_pc,
+                reads: ctx.reads.into(),
+                writes: ctx.writes.into(),
+                instruction: ctx.instruction.into(),
+            }
+        }
+    }
+
+    // AdapterRuntimeContext: DynInterface -> VecHeapAdapterInterface
+    impl<
+            T,
+            const R: usize,
+            const NUM_READS: usize,
+            const NUM_WRITES: usize,
+            const READ_SIZE: usize,
+            const WRITE_SIZE: usize,
+        > From<AdapterRuntimeContext<T, DynAdapterInterface<T>>>
+        for AdapterRuntimeContext<
+            T,
+            VecHeapAdapterInterface<T, R, NUM_READS, NUM_WRITES, READ_SIZE, WRITE_SIZE>,
+        >
+    {
+        fn from(ctx: AdapterRuntimeContext<T, DynAdapterInterface<T>>) -> Self {
+            AdapterRuntimeContext {
+                to_pc: ctx.to_pc,
+                writes: ctx.writes.into(),
+            }
+        }
+    }
 
     // AdapterAirContext: FlatInterface -> BasicInterface
     impl<
